requirements.txt
gradio>=3.50.0
sentence-transformers>=2.6.0
numpy
scikit-learn
tqdm
requests
gitpython
